Fix CodeQL implicit narrowing cast in LundLiteTest
Agent-Logs-Url: https://github.com/Antineutrino-4444/ColliderSimulator3/sessions/8e16acec-1976-4544-990a-e0418828dc43

+++ core/src/test/java/com/lhcsim/physics/collision/hadronize/LundLiteTest.java
@@ -59,7 +59,7 @@ class LundLiteTest {
             long charged = hadrons.stream()
                     .filter(h -> isCharged(h.pdgId()))
                     .count();
-            totalCharged += charged;
+            totalCharged += (int) charged;
         }
 
         double avgCharged = (double) totalCharged / nEvents;
